Updating the scripts to ensure accuracy of experiments.
The benchmarking script has its duration extended, and intervals get
changed to every 5 seconds, to avoid measurement inconsistancy due to
too high polling intervals. pkill is used instead of kill.
Adding a script to clean up old files
Adding a script to load the sql into the database
Updating the merge script so that an extra field shows which run the
current experiment is on
the tosql script makes an adjustment so that the correct index is used,
(given the new field)